requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.3.1
django-filter==23.2
django-jazzmin==2.6.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.51.0
frozenlist==1.4.1
idna==3.7
inference-sdk==0.9.22
inflection==0.5.1
kiwisolver==1.4.5
Markdown==3.4.3
marshmallow==3.21.1
matplotlib==3.8.4
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.1
pillow==10.3.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
supervision==0.19.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
